Deleted unnecessary void_stats_helpers module
